+++ prometheus_network_exporter/devices/basedevice.py
