+++ prometheus_network_exporter/views/junos/environment.yml
