+++ prometheus_network_exporter/views/junos/interface_metrics.yml
