+++ prometheus_network_exporter/views/junos/optic.yml
