+++ prometheus_network_exporter/views/junos/ospf.yml
